build: add support for requiring ejs files
Those files are using template() to be built, and will be built in place and
released on npm. You can require those templates if you need, but they need
to have a .ejs extension.

This required embedding the escape function so that templates' source
can be executed without any arguments. Additionally, also added default
options to the template function.