feat(@angular-devkit/schematics): support -- args
Passing -- on the command line will continue parsing arguments after
the --, but does not distinguish with arguments that are supported
by schematics directly.

e.g.  "schematics test --arg1=abc -- --dry-run=hello"
will have two arguments [ args, dry-run ]. If the -- was not passed
in dry-run would be removed from the options to the schematics.