feat(@angular-devkit/schematics): add support for aliases
Aliases are supported in the FileSystemEngineHostBase (and derivatives). When a
schematic is resolved the name being resolved will be replaced if a schematic
of the collection has an alias that matches the name.

Having 2 aliases be the same on two different schematics, or having an alias
be the same name as a schematic are both an error.

Also added much needed tests for file-system-engine-host.